move unit tests to their respective seperate files
