+++ src/test/java/com/brandonlenz/iso8583/messages/Iso8583MessageTest.java
@@ -6,8 +6,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import com.brandonlenz.iso8583.definitions.messages.Iso8583MessageDefinition;
 import com.brandonlenz.iso8583.definitions.messages.SampleIso8583MessageDefinition;
 import com.brandonlenz.iso8583.fields.DataField;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 class Iso8583MessageTest {
@@ -31,7 +29,6 @@ class Iso8583MessageTest {
     private final DataField messageTypeIndicator = messageDefinition.getMessageTypeIndicatorDefinition()
             .getDataFieldBuilder()
             .build("0100");
-    private Iso8583MessageBuilder iso8583MessageBuilder;
     private final Iso8583Message iso8583Message = getTestMessage();
 
     private Iso8583Message getTestMessage() {
@@ -44,67 +41,6 @@ class Iso8583MessageTest {
                 .setField(ICC_DATA_FIELD_NUMBER, ICC_DATA_FIELD_DATA)
                 .setField(SETTLEMENT_CODE_FIELD_NUMBER, SETTLEMENT_CODE_FIELD_DATA)
                 .build();
-    }
-
-    @BeforeEach
-    void setUp() {
-        this.iso8583MessageBuilder = Iso8583Message.builder(messageDefinition, messageTypeIndicator);
-    }
-
-    @Test
-    void setFieldString() {
-        Iso8583Message iso8583Message = iso8583MessageBuilder
-                .setField(AMOUNT_FIELD_NUMBER, AMOUNT_FIELD_DATA)
-                .build();
-
-        Assertions.assertTrue(iso8583Message.dataFieldIsSet(AMOUNT_FIELD_NUMBER));
-        Assertions.assertTrue(iso8583Message.dataFieldBitIsSet(AMOUNT_FIELD_NUMBER));
-    }
-
-    @Test
-    void setFieldBytes() {
-        Iso8583Message iso8583Message = iso8583MessageBuilder
-                .setField(AMOUNT_FIELD_NUMBER, AMOUNT_FIELD_RAW_DATA)
-                .build();
-
-        Assertions.assertTrue(iso8583Message.dataFieldIsSet(AMOUNT_FIELD_NUMBER));
-        Assertions.assertTrue(iso8583Message.dataFieldBitIsSet(AMOUNT_FIELD_NUMBER));
-    }
-
-    @Test
-    void removeField() {
-        Iso8583Message iso8583Message = iso8583MessageBuilder
-                .setField(AMOUNT_FIELD_NUMBER, AMOUNT_FIELD_DATA)
-                .removeField(AMOUNT_FIELD_NUMBER)
-                .build();
-
-        Assertions.assertFalse(iso8583Message.dataFieldIsSet(AMOUNT_FIELD_NUMBER));
-        Assertions.assertFalse(iso8583Message.dataFieldBitIsSet(AMOUNT_FIELD_NUMBER));
-    }
-
-    @Test
-    void setFieldGovernedBySecondaryBitmap() {
-        Iso8583Message iso8583Message = iso8583MessageBuilder
-                .setField(SETTLEMENT_CODE_FIELD_NUMBER, SETTLEMENT_CODE_FIELD_DATA)
-                .build();
-
-        Assertions.assertTrue(iso8583Message.dataFieldIsSet(SETTLEMENT_CODE_FIELD_NUMBER));
-        Assertions.assertTrue(iso8583Message.dataFieldBitIsSet(SETTLEMENT_CODE_FIELD_NUMBER));
-        Assertions.assertTrue(iso8583Message.dataFieldIsSet(messageDefinition.getSecondaryBitmapFieldNumber()));
-        Assertions.assertTrue(iso8583Message.dataFieldBitIsSet(messageDefinition.getSecondaryBitmapFieldNumber()));
-    }
-
-    @Test
-    void removeFieldGovernedBySecondaryBitmap() {
-        Iso8583Message iso8583Message = iso8583MessageBuilder
-                .setField(SETTLEMENT_CODE_FIELD_NUMBER, SETTLEMENT_CODE_FIELD_DATA)
-                .removeField(SETTLEMENT_CODE_FIELD_NUMBER)
-                .build();
-
-        Assertions.assertFalse(iso8583Message.dataFieldIsSet(SETTLEMENT_CODE_FIELD_NUMBER));
-        Assertions.assertFalse(iso8583Message.dataFieldIsSet(messageDefinition.getSecondaryBitmapFieldNumber()));
-        Assertions.assertFalse(iso8583Message.dataFieldBitIsSet(SETTLEMENT_CODE_FIELD_NUMBER));
-        Assertions.assertFalse(iso8583Message.dataFieldBitIsSet(messageDefinition.getSecondaryBitmapFieldNumber()));
     }
 
     @Test
